requirements.txt
azure-functions
azure-functions-durable==1.3.2
Pillow
pyodbc
requests